[UPDATED] Moved the example tests into their old folder. [ADDED] Created all the remaining tests with TODOs.
