+++ train-departures-common/src/commonTest/kotlin/uk/co/baconi/pka/common/openldbws/requests/examples/GetDepartureBoardRequestTest.kt
@@ -1,0 +1,5 @@
+package uk.co.baconi.pka.common.openldbws.requests.examples
+
+class GetDepartureBoardRequestTest {
+    // TODO
+}
