+++ train-departures-common/src/commonTest/kotlin/uk/co/baconi/pka/common/openldbws/services/CallingPointsTest.kt
@@ -1,0 +1,5 @@
+package uk.co.baconi.pka.common.openldbws.services
+
+class CallingPointsTest {
+    // TODO
+}
