+++ train-departures-common/src/commonTest/kotlin/uk/co/baconi/pka/common/openldbws/services/ToiletTest.kt
@@ -1,0 +1,5 @@
+package uk.co.baconi.pka.common.openldbws.services
+
+class ToiletTest {
+    // TODO
+}
